Split title, URL for requests from Android Chrome
Resolves an issue I (sole user!) encounter regularly: when I share a
link to HTTP Shortcuts from Chrome on Android, the shared text includes
both a URL and the title of the page --- apparently a behavior
introduced on Android a few months ago.

This change does some light munging to try to detect those cases: it
checks if the last token in the received data is a valid URL.

Adds a dependency on testy so I can unit-test the parsing.