Serve a bookmarklet script
The idea here is to support a "Send to pseudofeed" bookmarklet that
works like Feedbin's, by embedding a script in the current page and
executing it.

My initial bookmarklet implementation tried to execute a POST request
from bookmarklet Javascript, but ran into a common security restriction:
it violates `script-src 'unsafe-inline'`.

Feedbin violates `script-src [trusted-domain]`, but apparently that is
only loosely enforced.

I will revert this change if it doesn't work.